feat(vexdb-duck): CREATE INDEX option validation and runtime config range checks
Tighten input validation so bad inputs are rejected rather than silently
accepted (most of these caught existing test expectations from the main
branch test suite).

CREATE INDEX (vexdb-duck/index/graph_index.cpp):
  - Reject a non-vector first column (already checked) and any subsequent
    FLOAT[N] column (only the first column may be the vector).
  - Reject duplicate columns in the expression list. DuckDB collapses dups
    in column_ids, so the size mismatch with unbound_expressions is the
    detection signal.
  - 'm' must be in [2, 128].
  - 'ef_construction' must be in [1, 10000].
  - 'threads' must be a valid integer in [1, 1024]; option is accepted but
    ignored since vexdb-duck builds single-threaded.
  - 'pq_m' must be a non-negative integer; accepted but ignored (no PQ).
  - All other option names are rejected with "unknown option".
  - Cast errors for non-integer values produce uniform "must be a valid
    integer" messages.

Runtime config (vexdb-duck/optimizer/vex_physical_index_scan.cpp):
  - Validate vex_ef_search at search time: must be in [1, 65535].
  - Validate vex_brute_force_threshold at search time: [0, 1000000].

Defaults (vexdb-duck/vex_extension.cpp):
  - vex_ef_search default lowered from 64 to 40 to match the documentation
    in CLAUDE.md and the main-branch sqllogic test corpus.

+++ vexdb-duck/vex_extension.cpp
@@ -43,7 +43,7 @@ void LoadInternal(ExtensionLoader &loader) {
     OptimizerExtension::Register(config, VexOptimizerExtension());
 
     config.AddExtensionOption("vex_ef_search", "Search expansion factor for GRAPH_INDEX.",
-                              LogicalType::INTEGER, Value::INTEGER(64));
+                              LogicalType::INTEGER, Value::INTEGER(40));
     config.AddExtensionOption("vex_brute_force_threshold", "Temporary brute-force threshold for GRAPH_INDEX.",
                               LogicalType::UBIGINT, Value::UBIGINT(64));
 }
